requirements.txt
colorama
cycler
fonttools
fvcore
iopath
kiwisolver
matplotlib
numpy
opencv-contrib-python
packaging
Pillow
pip
portalocker
pyparsing
python-dateutil
PyYAML
scipy
setuptools
six
tabulate
termcolor
tqdm
trimesh
typing_extensions
wheel
yacs
imageio
pypose
numba
plyfile
torchgeometry